comps/retrievers/langchain/vdms/requirement.txt
# OPEA
beautifulsoup4
docarray[full]
fastapi
numpy
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
pandas
Pillow
prometheus_fastapi_instrumentator
shortuuid

# microservice
langchain==0.2.9
langchain-core==0.2.21
langchain-experimental==0.0.60
langchain-text-splitters==0.2.2
langsmith==0.1.92
open-clip-torch==2.24.0 # CLIP
vdms
faiss-cpu
decord
einops
torch==1.13.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.14.1 --index-url https://download.pytorch.org/whl/cpu
dateparser